requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pickle-mixin
requests
